pyproject.toml
[project]
name = "repl-nix-stockxscraper"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.3",
    "selenium-wire>=5.1.0",
    "trafilatura>=2.0.0",
    "undetected-chromedriver>=3.5.5",
    "webdriver-manager>=4.0.2",
]
